+++ packages/tsurlfilter/src/request-type.ts
@@ -1,29 +1,35 @@
 /**
- * RequestType is the request types enumeration
+ * RequestType is the request types enumeration.
+ * Important: the enumeration is marked as const to avoid side effects when
+ * importing it into an extension.
  */
-export enum RequestType {
+export const RequestType = {
     /** main frame */
-    Document = 1,
+    Document: 1,
     /** (iframe) $subdocument */
-    SubDocument = 1 << 1,
+    SubDocument: 2, // 1 << 1
     /** (javascript, etc) $script */
-    Script = 1 << 2,
+    Script: 4, // 1 << 2
     /** (css) $stylesheet */
-    Stylesheet = 1 << 3,
+    Stylesheet: 8, // 1 << 3
     /** (flash, etc) $object */
-    Object = 1 << 4,
+    Object: 16, // 1 << 4
     /** (any image) $image */
-    Image = 1 << 5,
+    Image: 32, // 1 << 5
     /** (ajax/fetch) $xmlhttprequest */
-    XmlHttpRequest = 1 << 6,
+    XmlHttpRequest: 64, // 1 << 6
     /** (video/music) $media */
-    Media = 1 << 7,
+    Media: 128, // 1 << 7
     /** (any custom font) $font */
-    Font = 1 << 8,
+    Font: 256, // 1 << 8
     /** (a websocket connection) $websocket */
-    WebSocket = 1 << 9,
+    WebSocket: 512, // 1 << 9
     /** (navigator.sendBeacon()) $ping */
-    Ping = 1 << 10,
+    Ping: 1024, // 1 << 10
     /** any other request type */
-    Other = 1 << 11,
-}
+    Other: 2048, // 1 << 11
+} as const;
+
+// intentionally naming the variable the same as the type
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export type RequestType = typeof RequestType[keyof typeof RequestType];