+++ packages/tswebextension/src/lib/mv2/content-script/assistant.ts
@@ -1,27 +1,28 @@
 import browser from 'webextension-polyfill';
-import { adguardAssistant, Assistant } from '@adguard/assistant';
-import { MessageType, sendAppMessage } from '../../common';
+// Import directly from files to avoid side effects of tree shaking.
+// If import from '../../common', entire tsurlfilter will be in the package.
+import { MessageType, sendAppMessage } from '../../common/content-script';
 
 /**
- * Initializes assistant object.
+ * Initializes assistant object and create messages listener for assistant.
  */
 export const initAssistant = (): void => {
     if (window.top !== window || !(document.documentElement instanceof HTMLElement)) {
         return;
     }
 
-    let assistant: Assistant;
-
-    browser.runtime.onMessage.addListener((message) => {
+    browser.runtime.onMessage.addListener(async (message) => {
         switch (message.type) {
             case MessageType.INIT_ASSISTANT: {
-                if (!assistant) {
-                    assistant = adguardAssistant();
+                // If there is no assistant on the window after execute
+                // loading script - throw error.
+                if (window.adguardAssistant === undefined) {
+                    throw new Error('adguardAssistant not found in the window object.');
                 } else {
-                    assistant.close();
+                    window.adguardAssistant.close();
                 }
 
-                assistant.start(null, (rules) => {
+                window.adguardAssistant.start(null, (rules) => {
                     sendAppMessage({
                         type: MessageType.ASSISTANT_CREATE_RULE,
                         payload: { ruleText: rules },
@@ -30,8 +31,8 @@ export const initAssistant = (): void => {
                 break;
             }
             case MessageType.CLOSE_ASSISTANT: {
-                if (assistant) {
-                    assistant.close();
+                if (window.adguardAssistant) {
+                    window.adguardAssistant.close();
                 }
                 break;
             }