+++ packages/tswebextension/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.57]
+
+### Changed
+- The order of imports to avoid side effects on tree shaking.
+- Made load of AdGuard Assistant lazy to decrease size of content-script bundle.
 
 ### Fixed
 - Fixed allowlist api rule generation and matching